+++ packages/crawlee-one/src/lib/io/dataset.ts
@@ -1,32 +1,8 @@
-import type { DatasetDataOptions, Log } from 'apify';
+import type { DatasetDataOptions } from 'apify';
 
 import { ValueMonitorOptions, createSizeMonitor } from '../../utils/valueMonitor.js';
 import type { CrawleeOneIO } from '../integrations/types.js';
 import { apifyIO } from '../integrations/apify.js';
-
-/**
- * Given a Dataset ID, get the number of entries already in the Dataset.
- *
- * By default uses Apify Dataset.
- */
-export const getDatasetCount = async (
-  datasetNameOrId?: string,
-  options?: { io?: CrawleeOneIO; log?: Log }
-) => {
-  const { io = apifyIO, log } = options ?? {};
-
-  log?.debug('Opening dataset');
-  const dataset = await io.openDataset(datasetNameOrId);
-  // const dataset = await io.openDataset(datasetNameOrId);
-  log?.debug('Obtaining dataset entries count');
-  const count = await dataset.getItemCount();
-  if (typeof count !== 'number') {
-    log?.warning('Failed to get count of entries in dataset. We use this info to know how many items were scraped. More entries might be scraped than was set.'); // prettier-ignore
-  } else {
-    log?.debug(`Done obtaining dataset entries count (${count})`);
-  }
-  return count;
-};
 
 /**
  * Given a Dataset ID and a name of a field, get the columnar data.