+++ packages/scraper-utils/src/index.ts
@@ -5,6 +5,9 @@ export * from './format.js';
 export * from './url.js';
 export * from './package.js';
 
+// Dataset utilities
+export * from './dataset.js';
+
 // Apify README rendering
 export * from './readme/apify/readme.js';
 export * from './readme/apify/types.js';